requirements.txt
https://www.djangoproject.com/download/1.7c1/tarball/#egg=Django
fabric
celery
django-celery
pep8
coverage
coveralls
djangorestframework
markdown
sphinx
mock
